requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
tavily-python>=0.3.0
flask>=3.0.0
